sync: Migrate timestamps to gint64
https://bugzilla.gnome.org/show_bug.cgi?id=790819

+++ lib/sync/ephy-synchronizable-manager.h
@@ -40,9 +40,9 @@ struct _EphySynchronizableManagerInterface {
   gboolean             (*is_initial_sync)         (EphySynchronizableManager *manager);
   void                 (*set_is_initial_sync)     (EphySynchronizableManager *manager,
                                                    gboolean                   is_initial);
-  double               (*get_sync_time)           (EphySynchronizableManager *manager);
+  gint64               (*get_sync_time)           (EphySynchronizableManager *manager);
   void                 (*set_sync_time)           (EphySynchronizableManager *manager,
-                                                   double                     sync_time);
+                                                   gint64                     sync_time);
   void                 (*add)                     (EphySynchronizableManager *manager,
                                                    EphySynchronizable        *synchronizable);
   void                 (*remove)                  (EphySynchronizableManager *manager,
@@ -62,9 +62,9 @@ GType               ephy_synchronizable_manager_get_synchronizable_type (EphySyn
 gboolean            ephy_synchronizable_manager_is_initial_sync         (EphySynchronizableManager *manager);
 void                ephy_synchronizable_manager_set_is_initial_sync     (EphySynchronizableManager *manager,
                                                                          gboolean                   is_initial);
-double              ephy_synchronizable_manager_get_sync_time           (EphySynchronizableManager *manager);
+gint64              ephy_synchronizable_manager_get_sync_time           (EphySynchronizableManager *manager);
 void                ephy_synchronizable_manager_set_sync_time           (EphySynchronizableManager *manager,
-                                                                         double                     sync_time);
+                                                                         gint64                     sync_time);
 void                ephy_synchronizable_manager_add                     (EphySynchronizableManager *manager,
                                                                          EphySynchronizable        *synchronizable);
 void                ephy_synchronizable_manager_remove                  (EphySynchronizableManager *manager,
